Rename manager_service.h as manager_service.hpp
