+++ worker/include/manager_service.hpp
