+++ worker/manager_service.cpp
@@ -7,17 +7,21 @@
 #include <exception>
 #include <iostream>
 
-#include "manager_service.h"
+#include "manager_service.hpp"
 
 using namespace boost;
 using boost::asio::ip::tcp;
 
 namespace ava_manager {
 
-ManagerServiceServerBase::ManagerServiceServerBase(uint32_t manager_port,
-        uint32_t worker_port_base, std::string worker_path, std::vector<std::string>& worker_argv) :
-    manager_port_(manager_port), worker_port_base_(worker_port_base),
-    worker_id_(0), worker_path_(worker_path), worker_argv_(worker_argv) {
+ManagerServiceServerBase::ManagerServiceServerBase(
+    uint32_t manager_port, uint32_t worker_port_base, std::string worker_path,
+    std::vector<std::string>& worker_argv)
+    : manager_port_(manager_port),
+      worker_port_base_(worker_port_base),
+      worker_id_(0),
+      worker_path_(worker_path),
+      worker_argv_(worker_argv) {
   acceptor_ = std::make_unique<tcp::acceptor>(
       io_service_, tcp::endpoint(tcp::v4(), manager_port));
   AcceptConnection();
@@ -88,7 +92,7 @@ ava_proto::WorkerAssignReply ManagerServiceServerBase::HandleRequest(
   parameters.push_back(std::to_string(port));
 
   // Append custom API server arguments
-  for (const auto &argv : worker_argv_) {
+  for (const auto& argv : worker_argv_) {
     parameters.push_back(argv);
   }
 
@@ -129,15 +133,15 @@ pid_t ManagerServiceServerBase::SpawnWorker(
   }
   envp_list.push_back(NULL);
 
-  std::vector<const char *> argv_list;
+  std::vector<const char*> argv_list;
   argv_list.push_back(worker_path_.c_str());
   for (auto& item : parameters) {
     argv_list.push_back(item.c_str());
   }
   argv_list.push_back(NULL);
 
-  if (execvpe(argv_list[0], (char* const*) argv_list.data(),
-        (char* const*)envp_list.data()) < 0)
+  if (execvpe(argv_list[0], (char* const*)argv_list.data(),
+              (char* const*)envp_list.data()) < 0)
     perror("execvpe worker failed");
 
   // Never reach here
